llm-service/pyproject.toml
[project]
name = "tarsy-llm"
version = "0.1.0"
description = "TARSy LLM Service - gRPC wrapper for Gemini client"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.68.1",
    "grpcio-tools>=1.68.1",
    "google-genai>=1.0.0",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["llm"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
